+++ scraper/utilities/parsers/pdb_strctre_parser.py
@@ -42,18 +42,35 @@
 
 
 import argparse
+import sys
 
 from pathlib import Path
 from typing import List, Optional
 
 
-def build_parser(argv: Optional[List] = None):
-    """Return ArgumentParser parser for script."""
+
+def build_pdb_structures_parser(argv: Optional[List] = None):
+    """Return ArgumentParser parser for the script 'expand.genbank_sequences.py'."""
     # Create parser object
     parser = argparse.ArgumentParser(
-        prog="cazy_webscraper.py",
-        description="Scrapes the CAZy database",
+        prog="pdb_structures.py",
+        description="Download structures from PDB",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+
+    # Add positional/required arguments
+    parser.add_argument(
+        "database",
+        type=Path,
+        metavar="local CAZy database",
+        help="Path to local CAZy database",
+    )
+
+    parser.add_argument(
+        "pdb",
+        choices=["mmCif", "pdb", "xml", "mmtf", "bundle"],
+        type=str,
+        help="File format of downloaded structure from PDB",
     )
 
     # Add optional arguments to parser
@@ -68,7 +85,7 @@ def build_parser(argv: Optional[List] = None):
         help="Path to configuration file. Default: None, scrapes entire database",
     )
 
-    # Add option to define complete classes to scrape
+    # Add option to define classes to retrieve protein sequences for
     parser.add_argument(
         "--classes",
         type=str,
@@ -76,50 +93,9 @@ def build_parser(argv: Optional[List] = None):
         help="Classes from which all families are to be scraped. Separate classes by ','"
     )
 
-    # Add option to use own CAZy class synoymn dict
+    # enable force writing in an existing directory
     parser.add_argument(
-        "--cazy_synonyms",
-        type=Path,
-        default=None,
-        help="Path to JSON file containing CAZy class synoymn names",
-    )
-
-    # Add option to display citation
-    parser.add_argument(
-        "-C",
-        "--citation",
-        dest="citation",
-        action="store_true",
-        default=False,
-        help="Print cazy_webscraper citation message",
-    )
-
-    parser.add_argument(
-        "--cache_dir",
-        type=Path,
-        default=None,
-        help="Target path for cache dir to be used instead of default path",
-    )
-
-    parser.add_argument(
-        "-d",
-        "--db_output",
-        type=Path,
-        default=None,
-        help="Target output path to build new SQL database",
-    )
-
-    parser.add_argument(
-        "-D",
-        "--database",
-        type=Path,
-        default=None,
-        help="Path to an existing local CAZy SQL database",
-    )
-
-    # Add option to force file over writting
-    parser.add_argument(
-        "-f",
+        "-f"
         "--force",
         dest="force",
         action="store_true",
@@ -127,12 +103,23 @@ def build_parser(argv: Optional[List] = None):
         help="Force file over writting",
     )
 
-    # Add option to specify families to scrape
+    # Add option to specify families to retrieve protein sequences for
     parser.add_argument(
         "--families",
         type=str,
         default=None,
-        help="Families to scrape. Separate families by commas 'GH1,GH2' (case sensitive)"
+        help="Families to scrape. Separate families by commas 'GH1,GH2'"
+    )
+
+    # Add option to restrict the scrape to specific kingdoms
+    parser.add_argument(
+        "--kingdoms",
+        type=str,
+        default=None,
+        help=(
+            "Kingdoms to scrape. Separate by a single comma.\n"
+            "Options= archaea, bacteria, eukaryota, viruses, unclassified (not case sensitive)"
+        ),
     )
 
     # Add option to restrict scrape to specific genera
@@ -154,49 +141,24 @@ def build_parser(argv: Optional[List] = None):
         help="Defines log file name and/or path",
     )
 
-    # Add option to not delete content in the existing cache dir
+    # Add option to prevent over writing of existing files
+    # and cause addition of files to output directory
     parser.add_argument(
-        "--nodelete_cache",
-        dest="nodelete_cache",
+        "-n",
+        "--nodelete",
+        dest="nodelete",
         action="store_true",
         default=False,
-        help="When called, content in the existing cache dir is NOT deleted",
+        help="enable/disable deletion of exisiting files",
     )
 
+    # enable specifying an output directory
     parser.add_argument(
-        "--nodelete_log",
-        dest="nodelete_log",
-        action="store_true",
-        default=False,
-        help="When called, content in the existing log dir is NOT deleted",
-    )
-
-    # Add option to enable number of times to retry scraping
-    parser.add_argument(
-        "-r",
-        "--retries",
-        type=int,
-        default=10,
-        help="Number of times to retry scraping a family or class page if error encountered",
-    )
-
-    # Add option to force file over writting
-    parser.add_argument(
-        "--sql_echo",
-        dest="sql_echo",
-        action="store_true",
-        default=False,
-        help="Set SQLite engine echo to True (SQLite will print its log messages)",
-    )
-
-    # Add option to enable retrieval of subfamilies
-    parser.add_argument(
-        "-s",
-        "--subfamilies",
-        dest="subfamilies",
-        action="store_true",
-        default=False,
-        help="Enable retrieval of subfamilies from CAZy",
+        "-o",
+        "--outdir",
+        type=Path,
+        metavar="output directory path",
+        help="Path to output directory to which downloaded structures are retrieved",
     )
 
     # Add option to restrict the scrape to specific species. This will scrape CAZymes from
@@ -219,26 +181,6 @@ def build_parser(argv: Optional[List] = None):
         ),
     )
 
-    # Add option to define time out limit for trying to connect to CAZy
-    parser.add_argument(
-        "-t",
-        "--timeout",
-        type=int,
-        default=45,
-        help="Connection timeout limit (seconds)"
-    )
-
-    parser.add_argument(
-        "--validate",
-        dest="validate",
-        action="store_true",
-        default=False,
-        help=(
-            "Retrieve CAZy fam population sizes from CAZy and use to check\n"
-            "the number of family members added to the local database"
-        ),
-    )
-
     # Add option for more detail (verbose) logging
     parser.add_argument(
         "-v",
@@ -248,16 +190,6 @@ def build_parser(argv: Optional[List] = None):
         default=False,
         help="Set logger level to 'INFO'",
     )
-    
-    # Add option to display version
-    parser.add_argument(
-        "-V",
-        "--version",
-        dest="version",
-        action="store_true",
-        default=False,
-        help="Print cazy_webscraper version number",
-    )    
 
     if argv is None:
         # parse command-line
